Moved dropdown functionality to dropdown.py.  Updated removal message to fit subreddit format.  Changed logger formatting.
